+++ app/login/layout.tsx
@@ -4,8 +4,12 @@ export default function LoginLayout({
     children: React.ReactNode
   }) {
     return (
-      <html lang="en">
-        <body>{children}</body>
+      <html lang="en" suppressHydrationWarning>
+        <body>
+          <div>
+              {children}
+          </div>
+        </body>
       </html>
     )
   }
