+++ app/signup/page.tsx
@@ -1,11 +1,22 @@
+"use client"
+
+import SignupModal from "@/src/components/signup-modal";
+import { useState } from "react";
 
 export default function SignupPage() {
+    
+    const [showSignupModal, setShowSignupModal] = useState(true);
+  
+    const closeSignupModal = () => {
+      setShowSignupModal(false);
+    };
+    
     return (
+        
       <section className="">
           <div>
-              <h1>
-                  sign up page
-              </h1>
+            <SignupModal showSignupModal={showSignupModal} closeSignupModal={closeSignupModal}/>
+
           </div>
       </section>
     )
