+++ src/components/login-modal.tsx
@@ -1,6 +1,5 @@
 import React, { useState } from "react";
 import styles from "@/src/styles/modal.module.css";
-import SignupModal from "./signup-modal";
 
 interface ModalProps {
   showModal: boolean;
@@ -19,30 +18,18 @@ const inputStyles = {
 
 }
 
-const Modal: React.FC<ModalProps> = ({ showModal, closeModal }) => {
+const SigninModal: React.FC<ModalProps> = ({ showModal, closeModal }) => {
 
     const handleBackdropClick = (event: React.MouseEvent<HTMLDivElement>) => {
         if (event.target === event.currentTarget) {
           closeModal();
+          window.location.href = '/';
         }
       };
 
     const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
         event.preventDefault();
         // logic TODO
-    };
-
-    const [showSignupModal, setShowSignupModal] = useState(false);
-
-    const handleSignupClick = (event: React.MouseEvent<HTMLAnchorElement>) => {
-      event.preventDefault();
-      
-      setShowSignupModal(true);
-      closeModal();
-    };
-  
-    const closeSignupModal = () => {
-      setShowSignupModal(false);
     };
 
   return (
@@ -70,16 +57,13 @@ const Modal: React.FC<ModalProps> = ({ showModal, closeModal }) => {
           </form>
           <div className={styles.signupSection}>
             <h1 className={styles.signUpText}> Not registered?</h1>
-            <a className={styles.signUpLink} href="#" onClick={handleSignupClick}> Sign up.</a>
-
+            <a className={styles.signUpLink} href="/signup"> Sign up.</a>
           </div>
         </div>
       </div>
     )}
-
-    <SignupModal showSignupModal={showSignupModal} closeSignupModal={closeSignupModal}/>
     </>
   );
 };
 
-export default Modal;   
+export default SigninModal;
